Move structs in Models & Remove useless clone
