+++ massa-consensus-worker/src/tests/scenarios_operations_check.rs
@@ -2,7 +2,7 @@
 use super::tools::*;
 use massa_consensus_exports::tools::*;
 
-use massa_models::ledger::LedgerData;
+use massa_models::ledger_models::LedgerData;
 use massa_models::prehash::Set;
 use massa_models::{Address, Amount, Slot};
 use massa_signature::{derive_public_key, generate_random_private_key, PrivateKey};
