+++ massa-models/src/error.rs
@@ -3,6 +3,8 @@
 use displaydoc::Display;
 use thiserror::Error;
 
+pub type ModelsResult<T, E = ModelsError> = core::result::Result<T, E>;
+
 #[non_exhaustive]
 #[derive(Display, Error, Debug)]
 pub enum ModelsError {
@@ -32,4 +34,8 @@ pub enum ModelsError {
     TimeOverflowError,
     /// Time error {0}
     TimeError(#[from] massa_time::TimeError),
+    /// invalid roll update: {0}
+    InvalidRollUpdate(String),
+    /// Ledger changes, Amount overflow
+    AmountOverflowError,
 }
