+++ massa-models/src/rolls.rs
@@ -1,12 +1,16 @@
+use serde::{Deserialize, Serialize};
 use std::collections::hash_map;
 
-use massa_models::{
+use crate::{
+    error::ModelsResult as Result,
     prehash::{Map, Set},
     Address, DeserializeCompact, DeserializeVarInt, ModelsError, SerializeCompact, SerializeVarInt,
 };
-use serde::{Deserialize, Serialize};
 
-use crate::{error::ProofOfStakeError, RollCompensation};
+use std::collections::{btree_map, BTreeMap};
+
+#[derive(Copy, Clone, Debug, Serialize, Deserialize)]
+pub struct RollCompensation(pub u64);
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct RollUpdate {
@@ -17,13 +21,13 @@ pub struct RollUpdate {
 
 impl RollUpdate {
     /// chain two roll updates, compensate and return compensation count
-    fn chain(&mut self, change: &Self) -> Result<RollCompensation, ProofOfStakeError> {
+    fn chain(&mut self, change: &Self) -> Result<RollCompensation> {
         let compensation_other = std::cmp::min(change.roll_purchases, change.roll_sales);
         self.roll_purchases = self
             .roll_purchases
             .checked_add(change.roll_purchases - compensation_other)
             .ok_or_else(|| {
-                ProofOfStakeError::InvalidRollUpdate(
+                ModelsError::InvalidRollUpdate(
                     "roll_purchases overflow in RollUpdate::chain".into(),
                 )
             })?;
@@ -31,9 +35,7 @@ impl RollUpdate {
             .roll_sales
             .checked_add(change.roll_sales - compensation_other)
             .ok_or_else(|| {
-                ProofOfStakeError::InvalidRollUpdate(
-                    "roll_sales overflow in RollUpdate::chain".into(),
-                )
+                ModelsError::InvalidRollUpdate("roll_sales overflow in RollUpdate::chain".into())
             })?;
 
         let compensation_self = self.compensate().0;
@@ -41,9 +43,7 @@ impl RollUpdate {
         let compensation_total = compensation_other
             .checked_add(compensation_self)
             .ok_or_else(|| {
-                ProofOfStakeError::InvalidRollUpdate(
-                    "compensation overflow in RollUpdate::chain".into(),
-                )
+                ModelsError::InvalidRollUpdate("compensation overflow in RollUpdate::chain".into())
             })?;
         Ok(RollCompensation(compensation_total))
     }
@@ -70,10 +70,7 @@ impl RollUpdates {
     }
 
     /// chains with another RollUpdates, compensates and returns compensations
-    pub fn chain(
-        &mut self,
-        updates: &RollUpdates,
-    ) -> Result<Map<Address, RollCompensation>, ProofOfStakeError> {
+    pub fn chain(&mut self, updates: &RollUpdates) -> Result<Map<Address, RollCompensation>> {
         let mut res = Map::default();
         for (addr, update) in updates.0.iter() {
             res.insert(*addr, self.apply(addr, update)?);
@@ -88,11 +85,7 @@ impl RollUpdates {
     }
 
     /// applies a RollUpdate, compensates and returns compensation
-    pub fn apply(
-        &mut self,
-        addr: &Address,
-        update: &RollUpdate,
-    ) -> Result<RollCompensation, ProofOfStakeError> {
+    pub fn apply(&mut self, addr: &Address, update: &RollUpdate) -> Result<RollCompensation> {
         if update.is_nil() {
             return Ok(RollCompensation(0));
         }
@@ -132,7 +125,7 @@ impl RollUpdates {
 }
 
 impl SerializeCompact for RollUpdate {
-    fn to_bytes_compact(&self) -> Result<Vec<u8>, ModelsError> {
+    fn to_bytes_compact(&self) -> Result<Vec<u8>> {
         let mut res: Vec<u8> = Vec::new();
 
         // roll purchases
@@ -146,7 +139,7 @@ impl SerializeCompact for RollUpdate {
 }
 
 impl DeserializeCompact for RollUpdate {
-    fn from_bytes_compact(buffer: &[u8]) -> Result<(Self, usize), ModelsError> {
+    fn from_bytes_compact(buffer: &[u8]) -> Result<(Self, usize)> {
         let mut cursor = 0usize;
 
         // roll purchases
@@ -164,5 +157,90 @@ impl DeserializeCompact for RollUpdate {
             },
             cursor,
         ))
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, Default)]
+pub struct RollCounts(pub BTreeMap<Address, u64>);
+
+impl RollCounts {
+    pub fn new() -> Self {
+        RollCounts(BTreeMap::new())
+    }
+
+    pub fn len(&self) -> usize {
+        self.0.len()
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.0.is_empty()
+    }
+
+    /// applies RollUpdates to self with compensations
+    pub fn apply_updates(&mut self, updates: &RollUpdates) -> Result<()> {
+        for (addr, update) in updates.0.iter() {
+            match self.0.entry(*addr) {
+                btree_map::Entry::Occupied(mut occ) => {
+                    let cur_val = *occ.get();
+                    if update.roll_purchases >= update.roll_sales {
+                        *occ.get_mut() = cur_val
+                            .checked_add(update.roll_purchases - update.roll_sales)
+                            .ok_or_else(|| {
+                                ModelsError::InvalidRollUpdate(
+                                    "overflow while incrementing roll count".into(),
+                                )
+                            })?;
+                    } else {
+                        *occ.get_mut() = cur_val
+                            .checked_sub(update.roll_sales - update.roll_purchases)
+                            .ok_or_else(|| {
+                                ModelsError::InvalidRollUpdate(
+                                    "underflow while decrementing roll count".into(),
+                                )
+                            })?;
+                    }
+                    if *occ.get() == 0 {
+                        // remove if 0
+                        occ.remove();
+                    }
+                }
+                btree_map::Entry::Vacant(vac) => {
+                    if update.roll_purchases >= update.roll_sales {
+                        if update.roll_purchases > update.roll_sales {
+                            // ignore if 0
+                            vac.insert(update.roll_purchases - update.roll_sales);
+                        }
+                    } else {
+                        return Err(ModelsError::InvalidRollUpdate(
+                            "underflow while decrementing roll count".into(),
+                        ));
+                    }
+                }
+            }
+        }
+        Ok(())
+    }
+
+    /// get roll counts for a subset of addresses.
+    #[must_use]
+    pub fn clone_subset(&self, addrs: &Set<Address>) -> Self {
+        Self(
+            addrs
+                .iter()
+                .filter_map(|addr| self.0.get(addr).map(|v| (*addr, *v)))
+                .collect(),
+        )
+    }
+
+    /// merge another roll counts into self, overwriting existing data
+    /// addrs that are in not other are removed from self
+    pub fn sync_from(&mut self, addrs: &Set<Address>, mut other: RollCounts) {
+        for addr in addrs.iter() {
+            if let Some(new_val) = other.0.remove(addr) {
+                self.0.insert(*addr, new_val);
+            } else {
+                self.0.remove(addr);
+            }
+        }
     }
 }
